Fix gravity, use timers and fix sync bug
[source/base.h]
- Add spawn height difference for players.

[source/client]
- Add Debug function to print position
- Use timers
- Remove Client::Tick()
- Draw based on timer instead of every loop
- Don't perform Thrust every frame, instead only put engine on when
  player presses up

[source/globals]
- Remove timekeeping globals

[source/player]
- Rename self_sustaining to is_local
- Fix broken physics logic

[source/session]
- Fix a bug in the ack mechanism that caused packets to be dropped

[source/timer]
- Use float type in timer

[source/types]
- Some tweaking of scaling
- Remove `REALITYSCALE`
- Remove `SHIPMASS` (assume a mass of 1)

+++ source/timer.cpp
@@ -1,6 +1,6 @@
 #include "timer.h"
 
-uint64_t Timer::delta = 0;
-uint64_t Timer::previous_tick_time = 0;
-uint64_t Timer::current_tick_time = 0;
+float Timer::delta = 0;
+float Timer::previous_tick_time = 0;
+float Timer::current_tick_time = 0;
 std::list<Timer*> Timer::instances;
